feat(JuniorChat): 채팅 상세 내용 관련 API 서비스 및 모델 분리
