+++ lib/features/junior/presentation/views/chat/junior_chat_screen.dart
@@ -15,140 +15,8 @@ import 'package:weve_client/core/constants/custom_svg_image.dart';
 import 'package:weve_client/core/constants/fonts.dart';
 import 'package:weve_client/core/enums/worry_status.dart';
 import 'package:weve_client/core/localization/app_localizations.dart';
-import 'package:weve_client/core/utils/api_client.dart';
 import 'package:flutter/foundation.dart';
-
-// 고민 상세 내용 응답 모델
-class WorryDetailResponse {
-  final bool isSuccess;
-  final String code;
-  final String message;
-  final WorryDetailResult? result;
-
-  WorryDetailResponse({
-    required this.isSuccess,
-    required this.code,
-    required this.message,
-    this.result,
-  });
-
-  factory WorryDetailResponse.fromJson(Map<String, dynamic> json) {
-    return WorryDetailResponse(
-      isSuccess: json['isSuccess'] as bool,
-      code: json['code'] as String,
-      message: json['message'] as String,
-      result: json['result'] != null
-          ? WorryDetailResult.fromJson(json['result'] as Map<String, dynamic>)
-          : null,
-    );
-  }
-}
-
-class WorryDetailResult {
-  final String content;
-  final String author;
-
-  WorryDetailResult({
-    required this.content,
-    required this.author,
-  });
-
-  factory WorryDetailResult.fromJson(Map<String, dynamic> json) {
-    return WorryDetailResult(
-      content: json['content'] as String,
-      author: json['author'] as String,
-    );
-  }
-}
-
-// 어르신 답변 응답 모델
-class SeniorAnswerResponse {
-  final bool isSuccess;
-  final String code;
-  final String message;
-  final SeniorAnswerResult? result;
-
-  SeniorAnswerResponse({
-    required this.isSuccess,
-    required this.code,
-    required this.message,
-    this.result,
-  });
-
-  factory SeniorAnswerResponse.fromJson(Map<String, dynamic> json) {
-    return SeniorAnswerResponse(
-      isSuccess: json['isSuccess'] as bool,
-      code: json['code'] as String,
-      message: json['message'] as String,
-      result: json['result'] != null
-          ? SeniorAnswerResult.fromJson(json['result'] as Map<String, dynamic>)
-          : null,
-    );
-  }
-}
-
-class SeniorAnswerResult {
-  final String content;
-  final String author;
-  final String imageUrl;
-
-  SeniorAnswerResult({
-    required this.content,
-    required this.author,
-    required this.imageUrl,
-  });
-
-  factory SeniorAnswerResult.fromJson(Map<String, dynamic> json) {
-    return SeniorAnswerResult(
-      content: json['content'] as String,
-      author: json['author'] as String,
-      imageUrl: json['imageUrl'] as String,
-    );
-  }
-}
-
-// 감사 인사 응답 모델
-class AppreciateResponse {
-  final bool isSuccess;
-  final String code;
-  final String message;
-  final AppreciateResult? result;
-
-  AppreciateResponse({
-    required this.isSuccess,
-    required this.code,
-    required this.message,
-    this.result,
-  });
-
-  factory AppreciateResponse.fromJson(Map<String, dynamic> json) {
-    return AppreciateResponse(
-      isSuccess: json['isSuccess'] as bool,
-      code: json['code'] as String,
-      message: json['message'] as String,
-      result: json['result'] != null
-          ? AppreciateResult.fromJson(json['result'] as Map<String, dynamic>)
-          : null,
-    );
-  }
-}
-
-class AppreciateResult {
-  final String content;
-  final String author;
-
-  AppreciateResult({
-    required this.content,
-    required this.author,
-  });
-
-  factory AppreciateResult.fromJson(Map<String, dynamic> json) {
-    return AppreciateResult(
-      content: json['content'] as String,
-      author: json['author'] as String,
-    );
-  }
-}
+import 'package:weve_client/features/junior/data/datasources/chat_detail_service.dart';
 
 // 상태 enum 정의
 
@@ -170,9 +38,7 @@ class JuniorChatScreen extends ConsumerStatefulWidget {
 
 class _JuniorChatScreenState extends ConsumerState<JuniorChatScreen> {
   bool _isHeadingBack = false;
-  final ApiClient _apiClient = ApiClient();
-  bool _isLoading = false;
-  String? _errorMessage;
+  final ChatDetailService _chatDetailService = ChatDetailService();
 
   @override
   void initState() {
@@ -200,147 +66,12 @@ class _JuniorChatScreenState extends ConsumerState<JuniorChatScreen> {
     return false;
   }
 
-  // 주니어 고민 상세 조회 API
-  Future<WorryDetailResult?> fetchWorryDetail(int worryId) async {
-    try {
-      setState(() {
-        _isLoading = true;
-        _errorMessage = null;
-      });
-
-      final response = await _apiClient.get<Map<String, dynamic>>(
-        '/api/worries/$worryId/junior',
-      );
-
-      if (kDebugMode) {
-        print('주니어 고민 상세 조회 응답: ${response.toJson()}');
-      }
-
-      if (response.isSuccess && response.code == 'COMMON200') {
-        final result = response.result;
-        if (result != null) {
-          return WorryDetailResult.fromJson(result);
-        }
-      } else {
-        setState(() {
-          _errorMessage = response.message;
-        });
-      }
-
-      return null;
-    } catch (e) {
-      if (kDebugMode) {
-        print('주니어 고민 상세 조회 오류: $e');
-      }
-
-      setState(() {
-        _errorMessage = '서버 오류가 발생했습니다.';
-      });
-
-      return null;
-    } finally {
-      setState(() {
-        _isLoading = false;
-      });
-    }
-  }
-
-  // 어르신 답변 API 조회
-  Future<SeniorAnswerResult?> fetchSeniorAnswer(int worryId) async {
-    try {
-      setState(() {
-        _isLoading = true;
-        _errorMessage = null;
-      });
-
-      final response = await _apiClient.get<Map<String, dynamic>>(
-        '/api/worries/$worryId/answer/junior',
-      );
-
-      if (kDebugMode) {
-        print('어르신 답변 조회 응답: ${response.toJson()}');
-      }
-
-      if (response.isSuccess && response.code == 'COMMON200') {
-        final result = response.result;
-        if (result != null) {
-          return SeniorAnswerResult.fromJson(result);
-        }
-      } else {
-        setState(() {
-          _errorMessage = response.message;
-        });
-      }
-
-      return null;
-    } catch (e) {
-      if (kDebugMode) {
-        print('어르신 답변 조회 오류: $e');
-      }
-
-      setState(() {
-        _errorMessage = '서버 오류가 발생했습니다.';
-      });
-
-      return null;
-    } finally {
-      setState(() {
-        _isLoading = false;
-      });
-    }
-  }
-
-  // 감사 인사 API 조회
-  Future<AppreciateResult?> fetchAppreciate(int worryId) async {
-    try {
-      setState(() {
-        _isLoading = true;
-        _errorMessage = null;
-      });
-
-      final response = await _apiClient.get<Map<String, dynamic>>(
-        '/api/worries/$worryId/appreciate/junior',
-      );
-
-      if (kDebugMode) {
-        print('감사 인사 조회 응답: ${response.toJson()}');
-      }
-
-      if (response.isSuccess && response.code == 'COMMON200') {
-        final result = response.result;
-        if (result != null) {
-          return AppreciateResult.fromJson(result);
-        }
-      } else {
-        setState(() {
-          _errorMessage = response.message;
-        });
-      }
-
-      return null;
-    } catch (e) {
-      if (kDebugMode) {
-        print('감사 인사 조회 오류: $e');
-      }
-
-      setState(() {
-        _errorMessage = '서버 오류가 발생했습니다.';
-      });
-
-      return null;
-    } finally {
-      setState(() {
-        _isLoading = false;
-      });
-    }
-  }
-
   // 공통 팝업 표시 함수
   void _showContentPopup({required String content, required String author}) {
     ref.read(popupProvider.notifier).showPopup(
           Container(
             width: double.infinity,
-            padding: const EdgeInsets.all(20),
+            padding: const EdgeInsets.all(10),
             decoration: BoxDecoration(
               color: WeveColor.gray.gray8,
               borderRadius: BorderRadius.circular(8),
@@ -370,7 +101,7 @@ class _JuniorChatScreenState extends ConsumerState<JuniorChatScreen> {
     ref.read(popupProvider.notifier).showPopup(
           Container(
             width: double.infinity,
-            padding: const EdgeInsets.all(30),
+            padding: const EdgeInsets.all(50),
             decoration: BoxDecoration(
               color: WeveColor.gray.gray8,
               borderRadius: BorderRadius.circular(8),
@@ -391,48 +122,73 @@ class _JuniorChatScreenState extends ConsumerState<JuniorChatScreen> {
           ),
         );
 
-    // API로 고민 상세 내용 가져오기
-    final detailResult = await fetchWorryDetail(widget.worryId);
+    try {
+      // API로 고민 상세 내용 가져오기
+      final detailResult =
+          await _chatDetailService.fetchWorryDetail(widget.worryId);
 
-    if (detailResult != null) {
-      // 데이터를 받은 후 팝업 내용 업데이트
-      ref.read(popupProvider.notifier).showPopup(
-            Container(
-              width: double.infinity,
-              padding: const EdgeInsets.all(20),
-              decoration: BoxDecoration(
-                color: WeveColor.gray.gray8,
-                borderRadius: BorderRadius.circular(8),
+      if (detailResult != null) {
+        // 데이터를 받은 후 팝업 내용 업데이트
+        ref.read(popupProvider.notifier).showPopup(
+              Container(
+                width: double.infinity,
+                padding: const EdgeInsets.all(10),
+                decoration: BoxDecoration(
+                  color: WeveColor.gray.gray8,
+                  borderRadius: BorderRadius.circular(8),
+                ),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.end,
+                  children: [
+                    const SizedBox(height: 10),
+                    Text(
+                      detailResult.content,
+                      style: WeveText.body2(color: WeveColor.main.yellowText),
+                    ),
+                    const SizedBox(height: 30),
+                    Text(
+                      detailResult.author,
+                      style: WeveText.body3(color: WeveColor.main.yellowText),
+                      textAlign: TextAlign.right,
+                    ),
+                    const SizedBox(height: 10),
+                  ],
+                ),
               ),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.end,
-                children: [
-                  Text(
-                    detailResult.content,
-                    style: WeveText.body2(color: WeveColor.main.yellowText),
-                  ),
-                  const SizedBox(height: 30),
-                  Text(
-                    detailResult.author,
-                    style: WeveText.body3(color: WeveColor.main.yellowText),
-                    textAlign: TextAlign.right,
-                  ),
-                ],
+            );
+      } else {
+        // 데이터가 없는 경우
+        ref.read(popupProvider.notifier).showPopup(
+              Container(
+                width: double.infinity,
+                padding: const EdgeInsets.all(30),
+                decoration: BoxDecoration(
+                  color: WeveColor.gray.gray8,
+                  borderRadius: BorderRadius.circular(8),
+                ),
+                child: Text(
+                  "고민 내용을 불러올 수 없습니다.",
+                  style: WeveText.body2(color: WeveColor.main.yellowText),
+                ),
               ),
-            ),
-          );
-    } else if (_errorMessage != null) {
+            );
+      }
+    } catch (e) {
       // 에러 메시지 표시
+      if (kDebugMode) {
+        print('고민 상세 조회 오류: $e');
+      }
+
       ref.read(popupProvider.notifier).showPopup(
             Container(
               width: double.infinity,
-              padding: const EdgeInsets.all(20),
+              padding: const EdgeInsets.all(10),
               decoration: BoxDecoration(
                 color: WeveColor.gray.gray8,
                 borderRadius: BorderRadius.circular(8),
               ),
               child: Text(
-                _errorMessage!,
+                "서버 오류가 발생했습니다.",
                 style: WeveText.body2(color: WeveColor.main.yellowText),
               ),
             ),
@@ -467,38 +223,63 @@ class _JuniorChatScreenState extends ConsumerState<JuniorChatScreen> {
           ),
         );
 
-    // API로 어르신 답변 내용 가져오기
-    final answerResult = await fetchSeniorAnswer(widget.worryId);
+    try {
+      // API로 어르신 답변 내용 가져오기
+      final answerResult =
+          await _chatDetailService.fetchSeniorAnswer(widget.worryId);
 
-    if (answerResult != null) {
-      // 데이터를 받은 후 팝업 내용 업데이트
-      ref.read(popupProvider.notifier).showPopup(
-            Container(
-              width: double.infinity,
-              padding: const EdgeInsets.all(30),
-              decoration: BoxDecoration(
-                color: WeveColor.gray.gray8,
-                borderRadius: BorderRadius.circular(8),
+      if (answerResult != null) {
+        // 데이터를 받은 후 팝업 내용 업데이트
+        ref.read(popupProvider.notifier).showPopup(
+              Container(
+                width: double.infinity,
+                padding: const EdgeInsets.all(10),
+                decoration: BoxDecoration(
+                  color: WeveColor.gray.gray8,
+                  borderRadius: BorderRadius.circular(8),
+                ),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.end,
+                  children: [
+                    const SizedBox(height: 10),
+                    Text(
+                      answerResult.content,
+                      style: WeveText.body2(color: WeveColor.main.yellowText),
+                    ),
+                    const SizedBox(height: 30),
+                    Text(
+                      answerResult.author,
+                      style: WeveText.body3(color: WeveColor.main.yellowText),
+                      textAlign: TextAlign.right,
+                    ),
+                    const SizedBox(height: 10),
+                  ],
+                ),
               ),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.end,
-                children: [
-                  Text(
-                    answerResult.content,
-                    style: WeveText.body2(color: WeveColor.main.yellowText),
-                  ),
-                  const SizedBox(height: 30),
-                  Text(
-                    answerResult.author,
-                    style: WeveText.body3(color: WeveColor.main.yellowText),
-                    textAlign: TextAlign.right,
-                  ),
-                ],
+            );
+      } else {
+        // 데이터가 없는 경우
+        ref.read(popupProvider.notifier).showPopup(
+              Container(
+                width: double.infinity,
+                padding: const EdgeInsets.all(30),
+                decoration: BoxDecoration(
+                  color: WeveColor.gray.gray8,
+                  borderRadius: BorderRadius.circular(8),
+                ),
+                child: Text(
+                  "어르신의 답변을 불러올 수 없습니다.",
+                  style: WeveText.body2(color: WeveColor.main.yellowText),
+                ),
               ),
-            ),
-          );
-    } else if (_errorMessage != null) {
+            );
+      }
+    } catch (e) {
       // 에러 메시지 표시
+      if (kDebugMode) {
+        print('어르신 답변 조회 오류: $e');
+      }
+
       ref.read(popupProvider.notifier).showPopup(
             Container(
               width: double.infinity,
@@ -508,7 +289,7 @@ class _JuniorChatScreenState extends ConsumerState<JuniorChatScreen> {
                 borderRadius: BorderRadius.circular(8),
               ),
               child: Text(
-                _errorMessage!,
+                "서버 오류가 발생했습니다.",
                 style: WeveText.body2(color: WeveColor.main.yellowText),
               ),
             ),
@@ -543,38 +324,63 @@ class _JuniorChatScreenState extends ConsumerState<JuniorChatScreen> {
           ),
         );
 
-    // API로 감사 인사 내용 가져오기
-    final appreciateResult = await fetchAppreciate(widget.worryId);
+    try {
+      // API로 감사 인사 내용 가져오기
+      final appreciateResult =
+          await _chatDetailService.fetchAppreciate(widget.worryId);
 
-    if (appreciateResult != null) {
-      // 데이터를 받은 후 팝업 내용 업데이트
-      ref.read(popupProvider.notifier).showPopup(
-            Container(
-              width: double.infinity,
-              padding: const EdgeInsets.all(30),
-              decoration: BoxDecoration(
-                color: WeveColor.gray.gray8,
-                borderRadius: BorderRadius.circular(8),
+      if (appreciateResult != null) {
+        // 데이터를 받은 후 팝업 내용 업데이트
+        ref.read(popupProvider.notifier).showPopup(
+              Container(
+                width: double.infinity,
+                padding: const EdgeInsets.all(10),
+                decoration: BoxDecoration(
+                  color: WeveColor.gray.gray8,
+                  borderRadius: BorderRadius.circular(8),
+                ),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.end,
+                  children: [
+                    const SizedBox(height: 10),
+                    Text(
+                      appreciateResult.content,
+                      style: WeveText.body2(color: WeveColor.main.yellowText),
+                    ),
+                    const SizedBox(height: 30),
+                    Text(
+                      appreciateResult.author,
+                      style: WeveText.body3(color: WeveColor.main.yellowText),
+                      textAlign: TextAlign.right,
+                    ),
+                    const SizedBox(height: 10),
+                  ],
+                ),
               ),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.end,
-                children: [
-                  Text(
-                    appreciateResult.content,
-                    style: WeveText.body2(color: WeveColor.main.yellowText),
-                  ),
-                  const SizedBox(height: 30),
-                  Text(
-                    appreciateResult.author,
-                    style: WeveText.body3(color: WeveColor.main.yellowText),
-                    textAlign: TextAlign.right,
-                  ),
-                ],
+            );
+      } else {
+        // 데이터가 없는 경우
+        ref.read(popupProvider.notifier).showPopup(
+              Container(
+                width: double.infinity,
+                padding: const EdgeInsets.all(30),
+                decoration: BoxDecoration(
+                  color: WeveColor.gray.gray8,
+                  borderRadius: BorderRadius.circular(8),
+                ),
+                child: Text(
+                  "감사 인사를 불러올 수 없습니다.",
+                  style: WeveText.body2(color: WeveColor.main.yellowText),
+                ),
               ),
-            ),
-          );
-    } else if (_errorMessage != null) {
+            );
+      }
+    } catch (e) {
       // 에러 메시지 표시
+      if (kDebugMode) {
+        print('감사 인사 조회 오류: $e');
+      }
+
       ref.read(popupProvider.notifier).showPopup(
             Container(
               width: double.infinity,
@@ -584,7 +390,7 @@ class _JuniorChatScreenState extends ConsumerState<JuniorChatScreen> {
                 borderRadius: BorderRadius.circular(8),
               ),
               child: Text(
-                _errorMessage!,
+                "서버 오류가 발생했습니다.",
                 style: WeveText.body2(color: WeveColor.main.yellowText),
               ),
             ),
